Add legacy session adapter class for reuse
